requirements.txt
pandas
numpy
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph
langchain-text-splitters
langchain_core
python-dotenv
docx2txt
ragas 
datasets
fastapi
pydantic
uvicorn 
jinja2 
python-multipart